agency properties added
